+++ web/src/App.tsx
@@ -10,6 +10,7 @@ import { ApiConfig } from "./components/ApiConfig";
 import { AnthropicSetup } from "./components/AnthropicSetup";
 import { CodeExamples } from "./components/CodeExamples";
 import { SettingsPanel } from "./components/SettingsPanel";
+import { QuotaSettings } from "./components/QuotaSettings";
 import { TestConnection } from "./components/TestConnection";
 import { Footer } from "./components/Footer";
 import { ProxySettings } from "./pages/ProxySettings";
@@ -148,6 +149,7 @@ function Dashboard() {
             serviceTier={status.selectedSpeed}
           />
           <SettingsPanel />
+          <QuotaSettings />
           <TestConnection />
         </div>
       </main>